feat: pre-build CLI interface
Add pre-built CLI interface
Minor model adjustments

+++ src/Exception/InsufficientBalanceException.java
@@ -1,16 +1,16 @@
 package Exception;
 
 public class InsufficientBalanceException extends RuntimeException {
-    private final int userBalance;
-    private final int requestedBalance;
+    private final long userBalance;
+    private final long requestedBalance;
 
-    public InsufficientBalanceException(int userBalance, int requestedBalance) {
+    public InsufficientBalanceException(long userBalance, long requestedBalance) {
         super("InsufficientBalanceException: userBalance=" + userBalance + " & requestedBalance=" + requestedBalance);
         this.userBalance = userBalance;
         this.requestedBalance = requestedBalance;
     }
 
-    public int getMoneyShortage() {
+    public long getMoneyShortage() {
         return requestedBalance - userBalance;
     }
 }
